add bind_interval
